search 컴포넌트 shared 폴더로 이동
