--- projects/han/src/app/city/search/search.component.html
@@ -1,2 +1,0 @@
-<input [formControl]="cityName" type="text" placeholder="도시 이름을 입력하세요">
-<button (click)="searchCity()">검색</button>
