+++ projects/han/src/app/shared/components/search/search.component.css
@@ -1,0 +1,12 @@
+.container {
+    display: flex;
+    margin: 50px 0 0 10px;
+}
+
+.searchBox {
+    order: 1;
+}
+
+.searchBtn {
+    order: 2;
+}
