+++ projects/han/src/app/shared/components/search/search.component.html
@@ -1,0 +1,9 @@
+<div class="container">
+    <input [formControl]="searchText" type="text" 
+        placeholder="{{ placeholderText }}" 
+        [style.width.px]="width" 
+        [style.height.px]="height"
+        class="searchBox">
+    <button (click)="submit()" 
+        class="searchBtn">검색</button>
+</div>
